AG-10859 - Ohlc options
