+++ packages/ag-charts-community/src/options/agChartOptions.ts
@@ -22,6 +22,7 @@ export * from './series/cartesian/cartesianSeriesTypes';
 export * from './series/cartesian/areaOptions';
 export * from './series/cartesian/barOptions';
 export * from './series/cartesian/boxPlotOptions';
+export * from './series/cartesian/candlestickBaseOptions';
 export * from './series/cartesian/candlestickOptions';
 export * from './series/cartesian/ohlcOptions';
 export * from './series/cartesian/bubbleOptions';
