+++ packages/ag-charts-community/src/options/series/cartesian/candlestickOptions.ts
@@ -2,56 +2,20 @@ import type { AgSeriesTooltip } from '../../chart/tooltipOptions';
 import type { CssColor } from '../../chart/types';
 import type { AgBaseCartesianThemeableOptions, AgBaseSeriesOptions } from '../seriesOptions';
 import type { AgBarSeriesStyle } from './barOptions';
-import type { AgCartesianSeriesTooltipRendererParams } from './cartesianSeriesTooltipOptions';
 import type {
-    AgSeriesFormatterParams,
-    AxisOptions,
-    FillOptions,
-    LineDashOptions,
-    StrokeOptions,
-} from './commonOptions';
-
-export type AgCandlestickSeriesItemType = 'up' | 'down';
-
-export interface AgCandlestickSeriesBaseOptions {
-    /** The key to use to retrieve open values from the data. */
-    openKey: string;
-    /** The key to use to retrieve close values from the data. */
-    closeKey: string;
-    /** The key to use to retrieve high values from the data. */
-    highKey: string;
-    /** The key to use to retrieve low values from the data. */
-    lowKey: string;
-    /** A human-readable description of open values. If supplied, this will be shown in the default tooltip and passed to the tooltip renderer as one of the parameters. */
-    openName?: string;
-    /** A human-readable description of close values. If supplied, this will be shown in the default tooltip and passed to the tooltip renderer as one of the parameters. */
-    closeName?: string;
-    /** A human-readable description of high values. If supplied, this will be shown in the default tooltip and passed to the tooltip renderer as one of the parameters. */
-    highName?: string;
-    /** A human-readable description of low values. If supplied, this will be shown in the default tooltip and passed to the tooltip renderer as one of the parameters. */
-    lowName?: string;
-}
+    AgCandlestickSeriesBaseFormatterParams,
+    AgCandlestickSeriesBaseOptions,
+    AgCandlestickSeriesBaseTooltipRendererParams,
+} from './candlestickBaseOptions';
+import type { AxisOptions, FillOptions, LineDashOptions, StrokeOptions } from './commonOptions';
 
 export type AgCandlestickWickOptions = StrokeOptions & LineDashOptions;
-
-export type AgCandlestickSeriesBaseFormatterParams<TDatum> = AgSeriesFormatterParams<TDatum> &
-    Readonly<
-        AgCandlestickSeriesBaseOptions &
-            Omit<AxisOptions, 'yKey'> &
-            StrokeOptions & {
-                /** Identifier showing whether the data element is rising (`up`) or falling (`down`). */
-                itemId: AgCandlestickSeriesItemType;
-                highlighted: boolean;
-            }
-    >;
 
 export interface AgCandlestickSeriesFormatterParams<TDatum>
     extends AgCandlestickSeriesBaseFormatterParams<TDatum>,
         Readonly<FillOptions> {}
 
-export interface AgCandlestickSeriesTooltipRendererParams
-    extends AgCandlestickSeriesBaseOptions,
-        Omit<AgCartesianSeriesTooltipRendererParams, 'yKey'> {
+export interface AgCandlestickSeriesTooltipRendererParams extends AgCandlestickSeriesBaseTooltipRendererParams {
     fill?: CssColor;
 }
